dss-plugin-visual-edit/tests/requirements.txt
behave
behave-pandas
dataiku-internal-client @ https://design.solutions.dataiku-dss.io/public/packages/dataiku-internal-client.tar.gz
dataiku-api-client
pandas==1.5.3
python-socketio==5.11.2
numpy==1.26.4
websocket-client
dssgherkin @ file:///Users/fabien.daoulas@dataiku.com/Documents/code/dss-gherkin-steps/dist/dssgherkin-0.0.1.tar.gz